requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
pytesseract
numpy
pillow
rembg[cpu]
onnxruntime